Repackaged at react. I'm a born again fan of brevity.
We package our artifact with our reverse domain, and if someone encounters a
transitive dependency on two libraries both packaged as react they can jarjar
their way to victory. The only remaining chance for conflict is if they really
want to use two libraries that both use react as a top-level package. The
chances of that are too vanishingly small to punish so many source files with
the repetition of com.threerings thousands of times.