SAMified Slot, added UnitSlot.
This feels a little nicer, in that one can use UnitSlot when they don't care
about the event type without even having to find out what the event type is and
declare it when they're instantiating the slot.

I didn't do the same for Value.Listener because there are three levels there
(value+ovalue, just value, and nothing). That starts to get ungainly, and I
couldn't think of elegant names for value+ovalue vs. just value.