Added a reactive Map implementation.
Note containsKeyView() and getView(), which are awesome. They will allow you
to do things like:

  gameObj.scores.getView(p1key).map(Functions.TO_STRING).
    connect(scoreLabel1.textSlot());

which will extract the score for player 1 from the scores map, convert it to a
string and bind it to a label in the user interface, such that when the score
changes in the distributed object, the label automatically changes.