Implemented a strategy for handling listener failures.
We want to ensure that regardless of whether a listener throws an unchecked
exception, it is properly disconnected if it is oneShot and that other
listeners are notified.

Rather than wire up some complex injectable logging mechanism (which will
introduce a world of pain when we try to make things work in GWT), I simply
collect up all failures that occur during a notification and throw them when
the notifications are complete.

This doesn't make life easier for users that would prefer to simply log any
notification failures and move on with their lives, but it does reduce the
collateral damage caused by a failing listener.