Add disconnect(Listener) methods
While Connection.disconnect accomplishes the same thing, it's a pain to
use in the case where a single listener is added to many dispatchers eg
a listener for clicks on multiple pieces of a display hierarchy. If
Connection is the only method to disconnect, the listener then has to
maintain a map from listened item to connection to disconnect if that
listened item becomes uninteresting. With this change, the listener can
just call disconnect with itself and not keep track of connections.

In the case where a Listener has been added to a Reactor multiple times,
disconnecting the listener disconnects all connections. If it's
disconnected during dispatch, the removal happens after dispatch
completes.

The one rough edge this introduces is that while you can still
Value.connect(Slot), there's not a corresponding disconnect(Slot). Since
the Slot is transformed into a Listener to connect to the Value, there's
no way for the user to pass in the actual Listener in the Cons.
I considered adding a toListener method on Slot to remove
Value.connect(Slot), but the fact that it'd expose a Listener where
nothing happened on no-arg onChange made it seem too gross.

This also fixes a couple minor bugs:

* Calling disconnect on a Connection to a Reactor with no listeners
  would throw an NPE
* listenNotify and connectNotify on Value dispatched before adding. That
  meant if the listener modified the value, it wouldn't be notified of
  its change.

Finally, I removed Signals.toggler rather than rework it for a world
where ValueView isn't a SignalView. I couldn't come up with a use case
for it, so I didn't see the worth in cluttering up the API with it. If
there's something I'm missing, I'll happily readd it.