Avoid leaking a connection if onEmit/onChange fail.
The reason we called onEmit/onChange before connect/listen was to avoid this
leakage, but since we want the listener to be logically connected by the time
they are notified, we'll just clean up after them ourselves if they fail.

Previously it was impossible for a listener to clean up after themselves if
onEmit/onChange threw an unchecked exception, because the connection instance
they would use to do it would never be returned to them. Now it is
theoretically possible for them to clean up after themselves because they can
use the new disconnect() method, but I'd rather not rely on callers being so
savvy as to understand this little edge case.