Judicious application of Occam's razor.
I realized that I can simply have Slot<T> extend ValueView.Listener<T> so that
one could use a Slot directly on a Value without having all of this mapping
machinery that we had previously constructed.

This eliminates Value.asSignal, as there is no longer any need for it. Values
are values and Signals are signals, it's Slots that can bridge the gap. If you
need to know the previous value, connect a ValueView.Listener to your Value. If
you don't, connect a Slot to your value. If you don't even care about the
value, connect a UnitSlot to your value. They are all the same.