Various experimental changes to make using React from Scala pleasant.
I'm parking these on this branch for now because there are a bunch of issues
that I can't properly resolve unless I was making heavy use of React in Scala,
which I'm not. Since React itself isn't changing much, these can wait and
Charlie can experiment with them locally in Papercut to see whether they are
moving in the right direction or whether a different approach is preferable.