initCause on triggering single MultiFailures
This gets a reasonable stack trace out of loggers and test runners that
generate the printout themselves from getStackTrace.

This is only filling in the trace for single non-Runtime exceptions, so
the trace will still suck for multi-exceptions from these tools. I first
tried overriding getStackTrace to return a big ol' array with all of the
StackTraceElements for all of the failures, but jamming non-contiguous
stacks together felt evil.