Fix inverted priority notification.
The old Connection.atPriority() is deprecated and inverts the supplied priority
before calling the new atPrio(). Connections are now notified from highest to
lowest priority, and in theory everything should just work out.

The only niggle is that MIN_VALUE is mapped to MAX_VALUE (because it cannot be
mapped to -MIN_VALUE), which results in it mapping to the same thing that
MIN_VALUE-1 maps to. It seems unlikely that anyone is listening at MIN_VALUE
and MIN_VALUE-1, but if they are, tough cookies.