Defer notify during notify instead of choking.
This introduces a bit more implementation complexity, but it makes life a lot
easier for clients.

In order to properly honor Connection.once() this also changes some edge case
behavior slightly. Previously, if a signal earlier in the dispatch chain
removed a signal later in the dispatch chain, that remove didn't take effect
until the current dispatch was complete. So the later signal still got that
event.

Now, because a signal earlier in the notification chain dispatch another event
during dispatch, that second event would get queued up for processing before
the removal which is queued up for processing after a once() signal fires. This
would cause the once signal to fire twice, which kind of violates the spirit of
the term "once".

So we immediately defang the Connection when it is disconnected, replacing its
listener ref with a ref to the NOOP listener, even before we actually remove
the Connection from the listener list. That way if a once Connection is
notified after it has been disconnected, it won't pass the notification on to
the original listener.

But this means that even in non-double-dispatch, if an earlier listener
disconnects a later listener, that listener is immediately defang and doesn't
participate in the current event dispatch. Given that any of this behavior
wasn't specified in the docs anywhere, I don't mind changing one edge case into
another, but there may well be poorly considered code out there in the wild
which will now break because it was living life on the edge. Such is life.

We could preserve the old behavior by first going through the listener list and
disconnecting any once() listeners and then doing normal dispatch, but I'm not
going to pessimize performance of this very low-level, performance sensitive
code just to preserve some unspecified behavior that was pretty rare to begin
with.

I also changed Reactor.insert() to Reactor.append() because it appends to the
pending runs list, and calling it insert is misleading.