Changes to support use of Java 8 lambdas.
Both Signal and Value can now be listened to via lambdas:

  Value<A> v = ...;
  v.connect((newV, oldV) -> ...);
  v.connect(newV -> ...);

  Signal<A> s = ...;
  s.connect(v -> ...);

This is not implemented in a way that I consider absolutely ideal, because I
want to preserve Java 1.7 compatibility for now. But the API is structured such
that when we do drop Java 1.7 support, I can introduce a default method on
SignalView.Listener and eliminate some "boxing" and everything will then be
roses and unicorns.